Trabajos de enfoque/Sistemas de big data/superfresh_bigdata/requirements.txt
pandas==2.2.2
numpy==1.26.4
scikit-learn==1.5.0
scipy==1.13.1
statsmodels==0.14.2
fastapi==0.111.0
uvicorn[standard]==0.30.1
streamlit==1.35.0
plotly==5.22.0
matplotlib==3.9.0
seaborn==0.13.2
requests==2.32.3
pydantic==2.7.1
# Procesamiento Big Data con Apache Spark
pyspark==3.5.1
# Soporte PostgreSQL (opcional)
psycopg2-binary==2.9.9
# Soporte MongoDB (opcional)
pymongo==4.7.2
